Consistently use UTF-16LE encoding for exported files
When a mail with such an attachment is opened in Thunderbird 60, the
words are shown in Chinese since Thunderbird messes up the encoding.
That's a bug in Thunderbird. Among others, it interprets the byte
sequence 0x0D 0x00 0x0A 0x00 as the code point U+0A0D, which doesn't
make sense at all. In Thunderbird 68 from December 2019 this problem is
fixed.

Other programs like Microsoft Excel are similarly broken. If the
encoding were declared as UTF-8 instead, Excel would not know that the
tab is the field separator and would merge all columns into one. This
really depends on the encoding of the file. With UTF-16LE, Excel can
handle the tab as field separator.